+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [0.2.0] Release 0.2 (2017-12-24)
+
+- Feature: Remove all folders
+- Feature: Warning for duplicated movies
+- Fix: Progress bar Hidden when not scanning
+
 ## [0.1.2]  Release 3 (2017-12-23)
 
 **Added**